fix: target browsers supporting BigInt
fix: client-side routing in static builds
