+++ package.json
@@ -31,9 +31,11 @@
   },
   "browserslist": {
     "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
+      "chrome >= 67",
+      "edge >= 79",
+      "firefox >= 68",
+      "opera >= 54",
+      "safari >= 14"
     ],
     "development": [
       "last 1 chrome version",
@@ -42,7 +44,6 @@
     ]
   },
   "devDependencies": {
-    "@tailwindcss/forms": "^0.5.0",
     "autoprefixer": "^10.4.5",
     "postcss": "^8.4.13",
     "tailwindcss": "^3.0.24"
